+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/RecordQosServiceImpl.java
@@ -11,7 +11,6 @@ import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.stack.bean.Bean;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
-import com.dwarfeng.subgrade.stack.exception.HandlerException;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
 import com.dwarfeng.subgrade.stack.log.LogLevel;
@@ -90,7 +89,7 @@ public class RecordQosServiceImpl implements RecordQosService {
     public RecordContext getRecordContext(LongIdKey pointKey) throws ServiceException {
         try {
             return recordLocalCacheHandler.get(pointKey);
-        } catch (HandlerException e) {
+        } catch (Exception e) {
             throw ServiceExceptionHelper.logParse("从本地缓存中获取记录上下文时发生异常", LogLevel.WARN, e, sem);
         }
     }